requirements.txt
# LangChain Core
langchain
langchain-core
langchain-experimental

# LangChain Community
langchain_community
pyPDF

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Groq-Integration
langchain-groq

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# PyTorch for local model execution
torch
torchvision
torchaudio

# sentence transformer for local
sentence-transformers

# Ui framework
streamlit

#Data validation
pydantic
pydantic[email]

# Beautifulsoup for web scraping
beautifulsoup4